TWINS-282: Refactor storager services for streamlined abstraction.
Consolidated redundant methods and classes across storager services to improve maintainability. Removed `StoragerLocalStatic`, updated logic for `getFileControllerUri`, and centralized handling of file URIs in `Storager`. Updated database migration scripts to reflect refactored class names.
