+++ core/src/main/resources/db/migration_backlog/V1.3.158.01__TWINS-282_resource_files_featurer.sql
@@ -5,25 +5,18 @@ on conflict (id) do update set name=excluded.name,
 
 --Local controller
 insert into public.featurer(id, featurer_type_id, class, name, description)
-values (2901, 29, 'org.twins.core.featurer.storager.StoragerLocalControllerFileService',
-        'StoragerLocalControllerFileService',
+values (2901, 29, 'org.twins.core.featurer.storager.StoragerLocal',
+        'StoragerLocal',
         'Service to save files in local file system and return them in controller after that')
 on conflict (id) do update set name=excluded.name,
                                description=excluded.description;
 
 
---Local static resource
-insert into public.featurer(id, featurer_type_id, class, name, description)
-values (2902, 29, 'org.twins.core.featurer.storager.StoragerLocalStaticFileService', 'StoragerLocalStaticFileService',
-        'Service to save files in local file system and return them them as nginx static resource after that')
-on conflict (id) do update set name=excluded.name,
-                               description=excluded.description;
-
 
 --External Uri
 insert into public.featurer(id, featurer_type_id, class, name, description)
-values (2903, 29, 'org.twins.core.featurer.storager.StoragerExternalUriFileService',
-        'StoragerExternalUriFileService',
+values (2903, 29, 'org.twins.core.featurer.storager.StoragerExternalUri',
+        'StoragerExternalUri',
         'Service to keep and work with external uri')
 on conflict (id) do update set name=excluded.name,
                                description=excluded.description;
@@ -31,7 +24,7 @@ on conflict (id) do update set name=excluded.name,
 
 --S3
 insert into public.featurer(id, featurer_type_id, class, name, description)
-values (2904, 29, 'org.twins.core.featurer.storager.StoragerS3FileControllerService', 'StoragerS3FileControllerService',
+values (2904, 29, 'org.twins.core.featurer.storager.StoragerS3', 'StoragerS3',
         'Service to save files to S3 and serve in controller')
 on conflict (id) do update set name=excluded.name,
                                description=excluded.description;
